Neural_Network_Service/requirements.txt
asyncpg
grpcio
grpcio-tools
numpy
scipy
tensorflow